fix: rename sortSelect folder
