+++ resources/js/components/listItem/ListItem.jsx
@@ -9,7 +9,7 @@ const ListItem = ({ item }) => {
                 <span className='title'>{item.title}</span>
             </h3>}
             <div>
-                <span className='comments'>Comments: {item.comments}</span>
+                <span className='comments'>Comments: {item.comments} </span>
                 <span className='points'>Points: {item.points}</span>
             </div>
         </StyledListItem>
